+++ lib/src/websocket_exception.dart
@@ -1,0 +1,8 @@
+class WebSocketException implements Exception {
+  final String message;
+
+  WebSocketException(this.message);
+
+  @override
+  String toString() => "WebSocketException: $message";
+}